Revert "♻️ convert SVG to PNG"
