+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "eliancodes-frontend",
-  "version": "0.7.1",
+  "version": "0.7.0",
   "private": true,
   "type": "module",
   "scripts": {
@@ -30,7 +30,6 @@
     "@unocss/preset-typography": "^0.49.8",
     "@unocss/preset-wind": "^0.49.8",
     "@unocss/reset": "^0.49.8",
-    "convert-svg-to-png": "^0.6.4",
     "eslint": "^8.35.0",
     "eslint-plugin-astro": "^0.23.0",
     "eslint-plugin-jsx-a11y": "^6.7.1",
