+++ src/components/layout/BaseHead.astro
@@ -8,7 +8,7 @@ interface Props {
 const canonicalURL = new URL(Astro.url.pathname, Astro.site);
 
 if (Astro.props.socialImg === undefined) {
-  Astro.props.socialImg = new URL('/v1/generate/og/default.png', Astro.url);
+  Astro.props.socialImg = new URL('/v1/generate/og/default.svg', Astro.url);
 }
 
 const { title, description, socialImg } = Astro.props;
