+++ src/layouts/BlogPost.astro
@@ -4,7 +4,7 @@ import BlogSideBar from '../components/blog/BlogSideBar.astro';
 import BlogContent from '../components/blog/BlogContent.astro';
 import RecentBlogPosts from '../components/generic/RecentBlogPosts.astro';
 
-const socialImg = new URL(`/v1/generate/og/${Astro.params.slug}.png`, Astro.url);
+const socialImg = new URL(`/v1/generate/og/${Astro.params.slug}.svg`, Astro.url);
 
 const { content, headings } = Astro.props;
 const { title, description } = content;
