--- src/pages/v1/generate/og/[slug].png.ts
@@ -1,68 +1,0 @@
-import type { APIRoute } from 'astro';
-import { getCollection } from 'astro:content';
-import satori from 'satori';
-import { html as toReactElement } from 'satori-html';
-// this is stupid
-// @ts-ignore
-import pkg from 'convert-svg-to-png';
-const { convert } = pkg;
-
-const fontFile = await fetch(
-  'https://og-playground.vercel.app/inter-latin-ext-700-normal.woff'
-);
-const fontData: ArrayBuffer = await fontFile.arrayBuffer();
-
-const height = 630;
-const width = 1200;
-
-const posts = await getCollection('blog');
-
-export function getStaticPaths() {
-  return posts.map((post) => ({
-    params: { slug: post.slug },
-    props: { title: post.data.title, description: post.data.description },
-  }));
-}
-
-export const get: APIRoute = async ({ params, props }) => {
-  const title = props.title.trim() ?? 'Blogpost';
-  const description = props.description ?? null;
-  const html = toReactElement(`
-  <div style="background-color: white; display: flex; flex-direction: column; height: 100%; padding: 3rem; width: 100%">
-    <div style="display:flex; height: 100%; width: 100%; background-color: white; border: 6px solid black; border-radius: 0.5rem; padding: 2rem; filter: drop-shadow(6px 6px 0 rgb(0 0 0 / 1));">
-      <div style="display: flex; flex-direction: column; justify-content: space-between; width: 100%; filter: drop-shadow()">
-        <div style="display: flex; justify-content: space-between;">
-          <div style="display: flex; flex-direction: column; gap: 0.75rem;">
-            <p style="font-size: 48px;">ElianCodes</p>
-            <p style="font-size: 38px;">${title}</p>
-          </div>
-          <img src="https://www.elian.codes/_astro/elian_ZJca6K.jpg" width="200px" height="200px" style="border: 3px solid black; border-radius: 0.5rem;" />
-        </div>
-        <div style="display: flex;">
-          <p style="font-size: 24px;">${description}</p>
-        </div>
-      </div>
-    </div>
-  </div>
-  `);
-
-  const svg = await satori(html, {
-    fonts: [
-      {
-        name: 'Inter Latin',
-        data: fontData,
-        style: 'normal',
-      },
-    ],
-    height,
-    width,
-  });
-
-  const img = await convert(svg);
-
-  return new Response(img, {
-    headers: {
-      'content-type': 'image/png',
-    },
-  });
-};
